requirements.txt
PyYAML
feedparser
openai>=1.0.0
requests
beautifulsoup4